+++ lib/lifer/entry.rb
@@ -33,6 +33,25 @@ class Lifer::Entry
   #
   DEFAULT_DATE = Time.new(1900, 01, 01, 0, 0, 0, "+00:00")
 
+  # If a filename contains a date, we should expect it to be in the following
+  # format.
+  #
+  FILENAME_DATE_FORMAT = /^(\d{4}-\d{1,2}-\d{1,2})-/
+
+  # We expect frontmatter to be provided in the following format.
+  #
+  FRONTMATTER_REGEX = /^---\n(.*)---\n/m
+
+  # If tags are represented in YAML frontmatter as a string, they're split on
+  # commas and/or spaces.
+  #
+  TAG_DELIMITER_REGEX = /[,\s]+/
+
+  # We truncate anything that needs to be truncated (summaries, meta
+  # descriptions) at the following character count.
+  #
+  TRUNCATION_THRESHOLD = 120
+
   attr_reader :file, :collection
 
   class << self
@@ -41,18 +60,13 @@ class Lifer::Entry
     #
     # @param file [String] The absolute filename of an entry file.
     # @param collection [Lifer::Collection] The collection for the entry.
-    # @return [Lifer::Entry::HTML, Lifer::Entry::Markdown]
+    # @return [Lifer::Entry] An entry.
     def generate(file:, collection:)
       error!(file) unless File.exist?(file)
 
       if (new_entry = subclass_for(file)&.new(file:, collection:))
         Lifer.entry_manifest << new_entry
-
-        # FIXME: Tags currently depend on frontmatter, so we should only
-        #   generate them for frontmatter supporting files (Markdown files). I
-        #   think the way forward is to parse every entry type for potential
-        #   frontmatter. Is that weird?
-        new_entry.tags if new_entry.is_a?(Lifer::Entry::Markdown)
+        new_entry.tags
       end
 
       new_entry
@@ -102,8 +116,8 @@ class Lifer::Entry
     end
   end
 
-  # When a new Markdown entry is initialized we expect the file to already
-  # exist, and we expect to know which `Lifer::Collection` it belongs to.
+  # When a new entry is initialized we expect the file to already exist, and
+  # we expect to know which `Lifer::Collection` it belongs to.
   #
   # @param file [String] An absolute path to a file.
   # @param collection [Lifer::Collection] A collection.
@@ -113,6 +127,48 @@ class Lifer::Entry
     @collection = collection
   end
 
+  # Given the entry's frontmatter, we should be able to get a list of authors.
+  # We always prefer authors (as opposed to a singular author) because it makes
+  # handling both cases easier in the long run.
+  #
+  # The return value here is likely an author's name. Whether that's a full
+  # name, a first name, or a handle is up to the end user.
+  #
+  # @return [Array<String>] An array of authors's names.
+  def authors
+    Array(frontmatter[:author] || frontmatter[:authors]).compact
+  end
+
+  # This method returns the full text of the entry, only removing the
+  # frontmatter. It should not parse anything other than frontmatter.
+  #
+  # @return [String] The body of the entry.
+  def body
+    return full_text.strip unless frontmatter?
+
+    full_text.gsub(FRONTMATTER_REGEX, "").strip
+  end
+
+  # Since text files would only store dates as simple strings, it's nice to
+  # attempt to convert those into Ruby date or datetime objects.
+  #
+  # @return [Time] A Ruby representation of the date and time provided by the
+  #   entry frontmatter or filename.
+  def date
+    date_data = frontmatter[:date] || filename_date
+
+    case date_data
+    when Time then date_data
+    when String then DateTime.parse(date_data).to_time
+    else
+      Lifer::Message.log("entry.no_date_metadata", filename: file)
+      Lifer::Entry::DEFAULT_DATE
+    end
+  rescue ArgumentError => error
+    Lifer::Message.error("entry.date_error", filename: file, error:)
+    Lifer::Entry::DEFAULT_DATE
+  end
+
   def feedable?
     if (setting = self.class.include_in_feeds).nil?
       raise NotImplementedError,
@@ -120,6 +176,19 @@ class Lifer::Entry
     end
 
     setting
+  end
+
+  # Frontmatter is a widely supported YAML metadata block found at the top of
+  # text--often Markdown--files. We attempt to parse all entries for
+  # frontmatter.
+  #
+  # @return [Hash] A hash representation of the entry frontmatter.
+  def frontmatter
+    return {} unless frontmatter?
+
+    Lifer::Utilities.symbolize_keys(
+      YAML.load(full_text[FRONTMATTER_REGEX, 1], permitted_classes: [Time])
+    )
   end
 
   # The full text of the entry.
@@ -162,14 +231,29 @@ class Lifer::Entry
   # @return [String] The absolute URI path to the entry.
   def path = permalink(host: "/")
 
-  # Currently, this method does nothing because not all entry types support
-  # tags. This method is only implemented to help other parts of the system
-  # iterate over tags where ever necessary.
+  # If given a summary in the frontmatter of the entry, we can use this to
+  # provide a summary.
   #
-  # @fixme We need to support frontmatter to get tags.
-  # @return [Array<Lifer::Tag>]
-  def tags = []
+  # Since subclasses may have more sophisticated access to the document, they
+  # may override this method with their own distinct implementations.
+  ##
+  # @return [String] A summary of the entry.
+  def summary
+    return frontmatter[:summary] if frontmatter[:summary]
+  end
 
+  # Locates and returns all tags defined in the entry.
+  #
+  # @return [Array<Lifer::Tag>] The entry's tags.
+  def tags
+    @tags ||= candidate_tag_names
+      .map { Lifer::Tag.build_or_update(name: _1, entries: [self]) }
+  end
+
+  # Returns the title of the entry. Every entry subclass must implement this
+  # method so that builders have access to *some* kind of title for each entry.
+  #
+  # @return [String]
   def title
     raise NotImplementedError, I18n.t("shared.not_implemented_method")
   end
@@ -177,5 +261,27 @@ class Lifer::Entry
   def to_html
     raise NotImplementedError, I18n.t("shared.not_implemented_method")
   end
-end
 
+  private
+
+  # It is conventional for users to use spaces or commas to delimit tags in
+  # other systems, so let's support that. But let's also support YAML-style
+  # arrays.
+  #
+  # @return [Array<String>] An array of candidate tag names.
+  def candidate_tag_names
+    case frontmatter[:tags]
+    when Array then frontmatter[:tags].map(&:to_s)
+    when String then frontmatter[:tags].split(TAG_DELIMITER_REGEX)
+    else []
+    end.uniq
+  end
+
+  def filename_date
+    return unless file && File.basename(file).match?(FILENAME_DATE_FORMAT)
+
+    File.basename(file).match(FILENAME_DATE_FORMAT)[1]
+  end
+
+  def frontmatter? = (full_text && full_text.match?(FRONTMATTER_REGEX))
+end