+++ spec/lifer/entry/html_spec.rb
@@ -5,18 +5,6 @@ RSpec.describe Lifer::Entry::HTML do
 
   let(:entry) { described_class.generate collection:, file: }
 
-  describe "#date" do
-    subject { entry.date }
-
-    let(:collection) {
-      Lifer::Collection.generate name: "Collection",
-        directory: File.dirname(file)
-    }
-    let(:file) { temp_file "file.html" }
-
-    it { is_expected.to be_a Time }
-  end
-
   describe "#title" do
     subject { entry.title }
 
@@ -24,11 +12,22 @@ RSpec.describe Lifer::Entry::HTML do
       Lifer::Collection.generate name: "Collection",
         directory: File.dirname(file)
     }
-    let(:file) { temp_file "test.html" }
 
-    it "returns the output filename" do
-      with_stdout_silenced do
-        expect(subject).to eq "test"
+    context "when there is title frontmatter" do
+      let(:file) { temp_file "test.html", "---\ntitle: Title\n---" }
+
+      it "returns the title frontmatter" do
+        expect(subject).to eq "Title"
+      end
+    end
+
+    context "when there is no title frontmatter" do
+      let(:file) { temp_file "test.html", "no frontmatter" }
+
+      it "returns the output filename" do
+        with_stdout_silenced do
+          expect(subject).to eq "test"
+        end
       end
     end
   end